+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change log for Microsoft365DSC
 
+# UNRELEASED
+
+* M365DSCGraphShim
+  * Changed how an object is returned from Graph if both `value` and
+    `presentation` properties are present, this fixes an issue on resource
+    `IntuneDeviceConfigurationAdministrativeTemplatePolicyWindows10`
+
 # 1.26.826.1
 
 * AADRoleAssignmentScheduleRequest